#1229 - extract 6 tile-type delegates from TileEditorModal
Repeats PR #1236's pattern for the remaining tile types — splitting the
~1.7k-line TileEditorModal into one delegate per tile.type:

  - TileEditorText.svelte             text body textarea
  - TileEditorChart.svelte            chart type + chart-options launcher
  - TileEditorFilter.svelte           widget + dim/hier/level + cascade
  - TileEditorKpi.svelte              measure/format/comparison/sparkline
                                       /threshold colouring
  - TileEditorTableConditional.svelte conditional-formatting rule builder
  - TileEditorTableSparkline.svelte   trailing trend-column config

Each delegate takes its state via $bindable() props so the parent stays
the persistence boundary — handleSave still reads the parent's variables
unchanged. The parent owns cube metadata + opens overlay modals (chart
options); delegates are pure UI.

Component-specific styles (.checkbox, .cf-* family, .chart-opts-btn)
moved into the delegates that use them; Svelte's scoped CSS does not
reach child component templates, so duplicating beats relying on the
parent's stylesheet to leak. Generic .field/.btn baseline still comes
from app.css. .radio remains in the parent because it's still used by
the parent template.

Verified: svelte-check clean (0 warnings, previously 10 unused-selector
warnings), vitest 867/867 pass.

Closes the bulk of the parent-split portion of saiku#1229.
